Move primitive aliases and utilities to a module accessible client and server.
