+++ kie-wb-common-services/kie-wb-common-data-modeller-core/src/main/java/org/kie/workbench/common/services/datamodeller/core/impl/PropertyTypeFactoryImpl.java
@@ -26,14 +26,14 @@ import java.util.List;
 import org.kie.workbench.common.services.datamodeller.core.PropertyType;
 import org.kie.workbench.common.services.datamodeller.core.PropertyTypeFactory;
 
-import static org.kie.workbench.common.services.datamodel.backend.server.DataModelOracleUtilities.BOOLEAN;
-import static org.kie.workbench.common.services.datamodel.backend.server.DataModelOracleUtilities.BYTE;
-import static org.kie.workbench.common.services.datamodel.backend.server.DataModelOracleUtilities.CHAR;
-import static org.kie.workbench.common.services.datamodel.backend.server.DataModelOracleUtilities.DOUBLE;
-import static org.kie.workbench.common.services.datamodel.backend.server.DataModelOracleUtilities.FLOAT;
-import static org.kie.workbench.common.services.datamodel.backend.server.DataModelOracleUtilities.INT;
-import static org.kie.workbench.common.services.datamodel.backend.server.DataModelOracleUtilities.LONG;
-import static org.kie.workbench.common.services.datamodel.backend.server.DataModelOracleUtilities.SHORT;
+import static org.kie.workbench.common.services.datamodel.util.PrimitiveUtilities.BOOLEAN;
+import static org.kie.workbench.common.services.datamodel.util.PrimitiveUtilities.BYTE;
+import static org.kie.workbench.common.services.datamodel.util.PrimitiveUtilities.CHAR;
+import static org.kie.workbench.common.services.datamodel.util.PrimitiveUtilities.DOUBLE;
+import static org.kie.workbench.common.services.datamodel.util.PrimitiveUtilities.FLOAT;
+import static org.kie.workbench.common.services.datamodel.util.PrimitiveUtilities.INT;
+import static org.kie.workbench.common.services.datamodel.util.PrimitiveUtilities.LONG;
+import static org.kie.workbench.common.services.datamodel.util.PrimitiveUtilities.SHORT;
 
 public class PropertyTypeFactoryImpl implements PropertyTypeFactory {
