+++ kie-wb-common-services/kie-wb-common-datamodel/kie-wb-common-datamodel-api/pom.xml
@@ -72,6 +72,13 @@
       <artifactId>kie-wb-common-services-api</artifactId>
     </dependency>
 
+    <!-- Test dependencies -->
+    <dependency>
+      <groupId>org.assertj</groupId>
+      <artifactId>assertj-core</artifactId>
+      <scope>test</scope>
+    </dependency>
+
   </dependencies>
 
 </project>
